requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==3.0.1  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client


# Data Analysis
pandas==0.24.2
scipy==1.2.1
Cython==0.29.7
ripser==0.3.2
kmapper==1.2.0

#Django External
django-downloadview==1.9

